Sección 14 Autenticación de protección de rutas

+++ src/app/api/todos/route.ts
@@ -1,6 +1,7 @@
 import prisma from '@/lib/prisma';
 import { NextResponse, type NextRequest } from 'next/server'
 import * as yup from 'yup';
+import { getUserSessionServer } from '@/auth/auth/auth-actions';
 
 export async function GET(request: NextRequest) { 
   
@@ -24,11 +25,14 @@ const postSchema = yup.object({
 })
 
 export async function POST(request: NextRequest) {
+
+    const user = await getUserSessionServer();
+    if (!user) return NextResponse.json({ message: 'No estás autorizado' }, { status: 401 });
     
     try {
         
         const {complete, description} = await request.json();
-        const todo = await prisma.todo.create({ data: { description, complete } });
+        const todo = await prisma.todo.create({ data: { description, complete, userId: user.id } });
      
         return NextResponse.json(todo);
         
@@ -39,8 +43,12 @@ export async function POST(request: NextRequest) {
 }
 
 export async function DELETE(request: NextRequest) {
+
+    const user = await getUserSessionServer();
+    if (!user) return NextResponse.json({ message: 'No estás autorizado' }, { status: 401 });
+
     try {
-        await prisma.todo.deleteMany({ where: { complete: true } });
+        await prisma.todo.deleteMany({ where: { complete: true, userId: user.id } });   
         return NextResponse.json({ message: 'Borrados todos los elementos completados' });
         
     } catch (error) {
